# internal/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PlaylistLabs User Management API",
    "description": "REST API for PlaylistLabs User Management open-source panel.\nAllows third-party systems, and automation tools to manage **Playlists & Sync**, **Managed Users**, **Team Members / Collaborators**, and **API Tokens**.\n\n### Authentication\nAuthentication is performed strictly using an internal API Token (`plt_live_...`) in the canonical **`X-API-Token`** header:\n- Header: `X-API-Token: <your_token>`\n\n### Optional Token Password\nIf the API Token was configured with an optional password, include it in the **`X-Token-Password`** header:\n- Header: `X-Token-Password: <your_password>`\n\n### Anti-Brute-Force & IP Banning\nInvalid tokens or repeated incorrect token passwords trigger automatic IP rate tracking.\nExceeding the maximum allowed failed attempts within the security window causes an **immediate IP ban** (`403 Forbidden`). Once banned, subsequent requests from the IP are rejected with no further logging.\n",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Current server base API path"
    }
  ],
  "security": [
    {
      "ApiTokenAuth": []
    },
    {
      "ApiTokenAuth": [],
      "TokenPasswordAuth": []
    }
  ],
  "paths": {
...
    "/tokens": {
      "get": {
        "summary": "List API tokens",
        "description": "Lists active and inactive API tokens created by the authenticated user.",
        "operationId": "getTokens",
        "tags": [
          "API Tokens"
        ],
        "responses": {
          "200": {
            "description": "List of API tokens",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/APIToken"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "summary": "Create an API token",
        "description": "Generates a new API token (`plt_live_...`). Plaintext secret is returned only once upon creation.",
        "operationId": "createToken",
        "tags": [
          "API Tokens"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTokenRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API token generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIToken"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/tokens/{id}": {
      "put": {
        "summary": "Update an API token",
        "description": "Updates configuration parameters of an existing API token, such as its name, allowed IP addresses, restricted playlist IDs, or optional password.",
        "operationId": "updateToken",
        "tags": [
          "API Tokens"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "API token ID",
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "API token updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIToken"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "summary": "Revoke an API token",
        "description": "Permanently deletes and invalidates an API token.",
        "operationId": "deleteToken",
        "tags": [
          "API Tokens"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token revoked",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "API token revoked successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/tokens/{id}/toggle": {
      "put": {
        "summary": "Toggle API token active status",
        "description": "Enables or disables an API token without deleting it.",
        "operationId": "toggleToken",
        "tags": [
          "API Tokens"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "is_active"
                ],
                "properties": {
                  "is_active": {
                    "type": "boolean",
                    "example": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "is_active": {
                      "type": "boolean",
                      "example": false
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Token",
        "description": "PlaylistLabs internal API Token (format: plt_live_...)"
      },
      "TokenPasswordAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Token-Password",
        "description": "Optional password configured during API token generation (leave empty if no password was set)"
      }
    },
    "schemas": {
...
      "APIToken": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "admin_id": {
            "type": "integer",
            "example": 2
          },
          "name": {
            "type": "string",
            "example": "Zapier Integration"
          },
          "token_prefix": {
            "type": "string",
            "example": "plt_live_7a8f...bc12"
          },
          "token": {
            "type": "string",
            "description": "Full plaintext token. Returned ONLY on creation.",
            "example": "plt_live_7a8f9c10d2e3f4a5b6c7d8e9f0123456789abcdef0123456789abcdef0123456"
          },
          "has_password": {
            "type": "boolean",
            "example": true
          },
          "allowed_ips": {
            "type": "string",
            "description": "Optional comma-separated list of allowed IP addresses or CIDRs.",
            "example": "192.168.1.50, 10.0.0.0/24"
          },
          "allowed_playlist_ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Optional list of playlist IDs this token is allowed to access. Empty or omitted means unrestricted access across all playlists managed by creator.",
            "example": [
              1001,
              1002
            ]
          },
          "last_used_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "is_active": {
            "type": "boolean",
            "example": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateTokenRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Billing CRM Bot"
          },
          "password": {
            "type": "string",
            "description": "Optional password. If specified, clients must also provide this password in X-Token-Password.",
            "example": "SuperSecretTokenPassword!"
          },
          "allowed_ips": {
            "type": "string",
            "description": "Optional comma-separated list of allowed IP addresses or CIDRs. If empty, all IPs are allowed.",
            "example": "192.168.1.50, 10.0.0.0/24"
          },
          "allowed_playlist_ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Optional list of playlist IDs this token is allowed to access. Empty or omitted means unrestricted access across all playlists managed by creator.",
            "example": [
              1001,
              1002
            ]
          },
          "expires_in_days": {
            "type": "integer",
            "description": "Number of days until token expiration. Set 0 for never expires.",
            "example": 90
          }
        }
      },
      "UpdateTokenRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Updated CRM Bot"
          },
          "password": {
            "type": "string",
            "description": "New password for token access. Provide empty string or leave unchanged if not updating.",
            "example": "NewStrongPassword123!"
          },
          "remove_password": {
            "type": "boolean",
            "description": "Set to true to remove password protection entirely.",
            "example": false
          },
          "allowed_ips": {
            "type": "string",
            "description": "Comma-separated list of allowed IP addresses or CIDRs. Provide empty string to allow all IPs.",
            "example": "192.168.1.50, 10.0.0.0/24"
          },
          "allowed_playlist_ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "description": "List of playlist IDs this token is allowed to access. Provide empty array to allow all playlists managed by creator.",
            "example": [
              1001
            ]
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "invalid_request"
          },
          "message": {
            "type": "string",
            "example": "Invalid parameters supplied"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request parameters or body",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication failed, missing token, expired token, or invalid token password",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Access denied, insufficient permissions, API tokens disabled for this account, or IP address blocked",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Requested resource was not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}